Initiative Pass + dice
-Added initiative rules
-Added Dice logic

-next basic in game actions

+++ CB2 Battle/Assets/Scripts/TacticsMovement.cs
@@ -14,7 +14,7 @@ public class TacticsMovement : MonoBehaviourPunCallbacks
    Tile currentTile;
 
    public bool moving = false!;
-   public float initative = 0;
+   public int initative = 0;
    public bool finishedMove = false;
    public float jumpHeight = 2;
    public float moveSpeed = 4;
